+++ CMakeLists.txt
@@ -4,24 +4,24 @@ find_package(CUDA REQUIRED)
 SET(CUDA_SEPARABLE_COMPILATION ON)
 
 set( CUDA_HEADER_FILES
-    source_files/Point.h
-    source_files/RandomnessGenerator.h
-    source_files/Ray.h
-    source_files/RandomWalk.h
-    source_files/Boundary.h
-    source_files/header.h
+    src/Point.h
+    src/RNG.h
+    src/Ray.h
+    src/RandomWalk.h
+    src/Boundary.h
+    src/header.h
 )
 
 set( CUDA_SOURCE_FILES
-    source_files/RandomnessGenerator.cu
-    source_files/Ray.cu
-    source_files/Point.cu
-    source_files/Boundary.cu
+    src/RNG.cu
+    src/Ray.cu
+    src/Point.cu
+    src/Boundary.cu
 )
 
 
 cuda_add_executable(
-    randomWalk source_files/main.cu
+    randomWalk src/main.cu
     ${CUDA_SOURCE_FILES}
     ${CUDA_HEADER_FILES}
     )
