--- source_files/Ray.h
@@ -1,26 +1,0 @@
-#ifndef Ray_H
-#define Ray_H
-#include "RandomnessGenerator.h"
-
-class Ray
-{
-
-public:
-__device__ Ray(Point startingPoint){    // Sets the ray starting point
-        this->_currentPos.setCoordinates(startingPoint.getX(), startingPoint.getY(), startingPoint.getZ());
-    } 
-__device__ void setDirection(Point direction);  // Sets the direction of the ray 
-__device__ void setStep(float step);           // Sets the step of movement
-__device__ Point getCurrentPos() const;
-__device__ Point getDirection() const;
-__device__ Point getPrevPos() const;
-__device__ float getStep() const;
-__device__ void move(); // The point moves in the specified direction with the given step -The function relies on member attributes that getYou should set first-
-
-private:
-    Point _prevPos;
-    Point _currentPos;
-    Point _direction;
-    float _step;
-};
-#endif
